Moved dependency loader to a separate class.
